Fix: Refresh token 관련 및 자잘한 오류 해결
1. access token 재발급 요청 시 invalid refresh token 에러를
   수정하였습니다.

   auth service에서 validate refreshtoken 메소드에서 refresh token을
   찾아도 return null을 하는 현상이 있었습니다. 아마 구문이 실행되는
   순서에서 문제가 있었던 거 같아 if else문으로 확실하게 구분 후 일단
   에러는 사라졌습니다. 하지만 올바른 해결방법이 아니여서 후에 다시
   문제가 생길 시 다시 수정할 수 있습니다.

2. getArticle 호출 시 dead lock 현상이 발생하면서 server error가
   발생하던 현상을 수정하였습니다.

   getArticle 호출이 2번씩 일어나고 있었고, 그에 따라 첫번째 트랜잭션이
   완료되지 않은 상태에서 두번째 호출이 진행되어 데드락 현상이 발생한
   것으로 예상됩니다.

   getArticle 메소드를 하위 컴포넌트에서 쓸 수 있도록 provide 하였는데
   이떄 onGetArticle이 아니라 onGetArticle()로 적어서 처음 스크립트문
   읽으면서 자동으로 실행된게 아닌가 생각됩니다. 이를 onGetArticle로
   수정 후 문제가 해결되었습니다.

3. 대댓글을 달 때, 여러개의 대댓글 입력 창을 띄울 수 이었습니다. 이를
   1개의 입력창만 띄울 수 있도록 수정하였습니다. 대댓글을 입력하다가
   다른 대댓글창을 열면, 기존 대댓글의 내용은 삭제되고, 기존 대댓글창은
   접히고, 새로 클릭한 대댓글창이 열리게 됩니다.

+++ src/main/java/com/codein/service/AuthService.java
@@ -4,9 +4,8 @@ import com.codein.domain.auth.Tokens;
 import com.codein.domain.member.Member;
 import com.codein.error.exception.auth.InvalidRefreshTokenException;
 import com.codein.error.exception.auth.RefreshTokenNullException;
-import com.codein.repository.TokensRepository;
+import com.codein.repository.tokens.TokensRepository;
 import com.codein.repository.member.MemberRepository;
-import jakarta.servlet.http.Cookie;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseCookie;
@@ -23,25 +22,26 @@ public class AuthService {
 
     @Transactional
     public ArrayList<String> validateRefreshToken(String refreshToken) {
-        Tokens token = tokensRepository.findByRefreshToken(refreshToken)
-                .orElseThrow(InvalidRefreshTokenException::new);
-
-        Member member = memberRepository.findByRefreshToken(token.getRefreshToken());
-
-        tokensRepository.delete(token);
-
-        if(member == null){
+        Tokens token = tokensRepository.findByRefreshToken(refreshToken);
+        if(token == null){  // db에 일치하는 refresh token이 없는 경우 쿠키를 삭제해야하기 때문에 null을 보내 쿠키를 삭제 후 예외처리
             return null;
-        }
-        else {
-            Tokens newTokens = Tokens.builder()
-                    .member(member)
-                    .build();
+        } else {
 
-            ArrayList<String> tokens = new ArrayList<>();
-            tokens.add(newTokens.getRefreshToken());
-            tokens.add(newTokens.getAccessToken());
-            return tokens;
+            Member member = memberRepository.findByRefreshToken(token.getRefreshToken());
+            tokensRepository.delete(token);
+            if (member == null) { // db에 refresh token을 소유하고 있는 member가 존재하지 않는 경우 null을 보내 refresh token 삭제
+                return null;
+            } else {
+
+                Tokens newTokens = Tokens.builder()
+                        .member(member)
+                        .build();
+
+                ArrayList<String> tokens = new ArrayList<>();
+                tokens.add(newTokens.getRefreshToken());
+                tokens.add(newTokens.getAccessToken());
+                return tokens;
+            }
         }
     }
 
@@ -82,11 +82,13 @@ public class AuthService {
     }
 
     @Transactional
-    public void deleteTokens(String refreshToken) {
-        Tokens tokens = tokensRepository.findByRefreshToken(refreshToken)
-                .orElseThrow(RefreshTokenNullException::new);
-
-        tokensRepository.delete(tokens);
+    public void deleteTokensByRefreshToken(String refreshToken) {
+        Tokens tokens = tokensRepository.findByRefreshToken(refreshToken);
+        if(tokens != null){
+            tokensRepository.delete(tokens);
+        } else {
+            throw new InvalidRefreshTokenException();
+        }
     }
 
 }
